split each subcmd into its own file
